+++ Project/src/main/java/egovframework/example/file/service/FileService.java
@@ -11,6 +11,6 @@ public interface FileService {
     FileVO getProfileFileByMemberId(Long memberId); // 👈 회원 프로필 조회용
     
     // 7/11 민중 게시글삭제를위한 달려있는 모든 파일 삭제 기능
-    void deleteAllByTargetIdAndType(int targetId, String useType); 
+    void deleteAllByTargetIdAndType(Long targetId, String useType); 
  }
